Update LICENSE.txt
changed from Non-Profit Open Software License 3.0 back to MIT license as NPOSL-3.0 had hidden wording that would've forced anyone hosting my mod on a server to have to disclose the source code. (Even though no one can download the mod from the server, the license seemed to be worded that "use" is the same as "distribution". But then again, I am not a lawyer lol)